Extract Commands::AbstractFilter.
