+++ lib/cuprum/collections/commands.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'cuprum/collections'
+
+module Cuprum::Collections
+  # Namespace for abstract commands implementing collection functionality.
+  module Commands; end
+end
