+++ lib/cuprum/collections/reference/commands/filter.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
+require 'cuprum/collections/commands/abstract_filter'
 require 'cuprum/collections/constraints/ordering'
-require 'cuprum/collections/queries/parse'
 require 'cuprum/collections/reference/command'
 require 'cuprum/collections/reference/commands'
 require 'cuprum/collections/reference/query'
@@ -9,6 +9,8 @@ require 'cuprum/collections/reference/query'
 module Cuprum::Collections::Reference::Commands
   # Command for querying filtered, ordered data from a reference collection.
   class Filter < Cuprum::Collections::Reference::Command
+    include Cuprum::Collections::Commands::AbstractFilter
+
     # @!method call(limit: nil, offset: nil, order: nil, &block)
     #   Queries the collection for items matching the given conditions.
     #
@@ -86,8 +88,9 @@ module Cuprum::Collections::Reference::Commands
     #     the query and wraps the resulting array in a Hash, using the name of
     #     the collection as the key.
     #
-    #     @return [Hash{String, Array}] a hash with the collection name as key
-    #       and the matching items as value.
+    #     @return [Cuprum::Result<Hash{String, Array<Hash{String, Object}>}>] a
+    #       hash with the collection name as key and the matching items as
+    #       value.
 
     # @param collection_name [String, Symbol] The name of the collection.
     # @param data [Array<Hash>] The current data in the collection.
@@ -110,53 +113,8 @@ module Cuprum::Collections::Reference::Commands
 
     private
 
-    def build_query(criteria:, limit:, offset:, order:)
-      query = Cuprum::Collections::Reference::Query.new(data)
-      query = query.limit(limit)   if limit
-      query = query.offset(offset) if offset
-      query = query.order(order)   if order
-      query = query.where(criteria, strategy: :unsafe) unless criteria.empty?
-
-      success(query)
-    end
-
-    def parse_criteria(strategy:, where:, &block)
-      return [] if strategy.nil? && where.nil? && !block_given?
-
-      Cuprum::Collections::Queries::Parse.new.call(
-        strategy: strategy,
-        where:    where || block
-      )
-    end
-
-    def process( # rubocop:disable Metrics/MethodLength, Metrics/ParameterLists
-      limit:    nil,
-      offset:   nil,
-      order:    nil,
-      strategy: nil,
-      where:    nil,
-      &block
-    )
-      criteria = step do
-        parse_criteria(strategy: strategy, where: where, &block)
-      end
-
-      query = step do
-        build_query(
-          criteria: criteria,
-          limit:    limit,
-          offset:   offset,
-          order:    order
-        )
-      end
-
-      success(wrap_query(query))
-    end
-
-    def wrap_query(query)
-      return query.each unless options[:envelope]
-
-      { collection_name => query.to_a }
+    def build_query
+      Cuprum::Collections::Reference::Query.new(data)
     end
   end
 end
